requirements.txt
# Requirements for Production/CI
# Generated from requirements-local.txt + production dependencies
Authlib>=1.3.2
Faker==37.1.0
GitPython==3.1.44
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
PyYAML==6.0.2
Pygments==2.19.2
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
coverage==7.9.2
cycler==0.12.1
duckdb==1.4.3
entrypoints==0.4
et_xmlfile==2.0.0
favicon==0.7.0
fonttools>=4.61.0
gitdb==4.0.12
google-analytics-data>=0.18.0
google-auth>=2.40.0,<3.0.0
htbuilder==0.9.0
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.1.0
joblib==1.4.2
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
kaggle==1.7.4.2
kagglehub==0.3.11
kiwisolver==1.4.8
lightgbm>=4.6.0
llvmlite>=0.45.0,<0.47
lxml==5.3.2
markdownlit==0.0.7
matplotlib==3.10.1
mypy==1.17.0
mypy_extensions==1.1.0
narwhals==1.33.0
nltk==3.9.1
numba>=0.62.1
numpy>=1.24.3,<2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
playwright==1.57.0
plotly==6.0.1
pluggy==1.6.0
prettytable==3.16.0
prometheus_client==0.21.1
prophet==1.1.7
protobuf>=3.20,<6.0.0
psutil==7.0.0
psycopg2-binary==2.9.11
pyarrow==19.0.1
pydeck==0.9.1
pyecharts==2.0.8
pymdown-extensions>=10.16.1
pyparsing==3.2.3
pytest-cov==6.2.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
referencing==0.36.2
regex>=2025.2.10
requests>=2.32.4
rpds-py==0.24.0
scikit-learn==1.8.0
seaborn==0.13.2
setuptools==78.1.0
shap==0.47.1
simplejson==3.20.1
six==1.17.0
# sklearn-compat==0.1.3  # Removido: conflito com scikit-learn>=1.7 (nao usado no codigo)
slicer==0.0.8
smmap==5.0.2
soupsieve==2.6
st-annotated-text==4.0.2
st-theme==1.2.3
statsmodels==0.14.5
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
# 1.44.x tem bug 400 NoneType na rota /auth/login; 1.45+ corrige e expõe st.user
streamlit>=1.45.0,<2
tenacity==9.1.2
tensorflow==2.20.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-Pillow==10.2.0.20240822
types-cachetools==6.0.0.20250525
types-pycurl==7.45.6.20250309
types-requests==2.32.4.20250611
types-toml==0.10.8.20240310
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.1
wordcloud==1.9.4
xgboost==3.0.0
